# backend/gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway банковского приложения. Единая точка входа для клиентских запросов.",
        "title": "Gateway Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает список всех активных, замороженных и архивных счетов текущего пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Список счетов",
                "responses": {
                    "200": {
                        "description": "Список счетов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.AccountDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Создаёт новый банковский счёт для текущего пользователя (RUB, USD, EUR).\nДоступные типы: checking, savings, credit, deposit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Открыть счёт",
                "parameters": [
                    {
                        "description": "Тип счёта и валюта",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.OpenAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Счёт успешно создан",
                        "schema": {
                            "$ref": "#/definitions/schemas.AccountDTO"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры (валюта или тип)",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/{account_id}": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает подробную информацию о конкретном банковском счёте, включая баланс и статус.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Детали счёта",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о счёте",
                        "schema": {
                            "$ref": "#/definitions/schemas.AccountDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Счёт не найден",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Переводит счёт в статус 'closed'. Проводки по нему становятся невозможны.\nСчёт может быть закрыт только при нулевом балансе.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Закрыть счёт",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счёт успешно закрыт",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Нельзя закрыть счёт с ненулевым балансом",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Счёт не найден",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{account_id}/suspensions": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Временно блокирует любые расходные операции по счёту.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Приостановить обслуживание (Заморозка)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счёт успешно заморожен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Счёт не найден",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Снимает временную блокировку со счёта, восстанавливая возможность проведения операций.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Возобновить обслуживание (Разморозка)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID счёта",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Счёт успешно разморожен",
                        "schema": {
                            "$ref": "#/definitions/schemas.SuccessResponse"
                        }
                    },
                    "403": {
                        "description": "Нет прав на разморозку (если заморожен системой)",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "schemas.AccountDTO": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "Доступный остаток средств",
                    "type": "number",
                    "example": 15000.5
                },
                "created_at": {
                    "description": "Дата и время открытия счета (ISO 8601)",
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-10-27T10:00:00Z"
                },
                "currency": {
                    "description": "Валюта (ISO 4217)",
                    "type": "string",
                    "enum": [
                        "RUB",
                        "USD",
                        "EUR"
                    ],
                    "example": "RUB"
                },
                "id": {
                    "description": "Уникальный идентификатор счёта (UUID)",
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "description": "Текущее состояние счёта",
                    "type": "string",
                    "enum": [
                        "active",
                        "frozen",
                        "closed"
                    ],
                    "example": "active"
                },
                "type": {
                    "description": "Тип продукта",
                    "type": "string",
                    "enum": [
                        "checking",
                        "savings",
                        "credit",
                        "deposit"
                    ],
                    "example": "checking"
                },
                "user_id": {
                    "description": "ID владельца счёта",
                    "type": "string",
                    "format": "uuid",
                    "example": "user-uuid"
                }
            }
        },
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Подробное сообщение об ошибке для разработчика и пользователя",
                    "type": "string",
                    "example": "Неверный PIN-код или формат данных"
                }
            }
        },
...
        "schemas.OpenAccountRequest": {
            "type": "object",
            "required": [
                "currency",
                "type"
            ],
            "properties": {
                "currency": {
                    "description": "Валюта счёта (ISO 4217)",
                    "type": "string",
                    "enum": [
                        "RUB",
                        "USD",
                        "EUR"
                    ],
                    "example": "RUB"
                },
                "type": {
                    "description": "Тип счёта: checking (расчётный), savings (накопительный), credit (кредитный), deposit (вклад)",
                    "type": "string",
                    "enum": [
                        "checking",
                        "savings",
                        "credit",
                        "deposit"
                    ],
                    "example": "checking"
                }
            }
        },
...
        "schemas.SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Статус успеха (обычно 'success')",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OnboardingToken": {
            "description": "Токен регистрации, полученный на Шаге 0 (Start Onboarding)",
            "type": "apiKey",
            "name": "X-Onboarding-Token",
            "in": "header"
        },
        "SessionToken": {
            "description": "Сессионный токен, полученный при авторизации",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    },
...
}